+++ nodeserver/config.js
@@ -1,6 +1,6 @@
 const { InfluxDB } = require('@influxdata/influxdb-client');
-const envSwitcher = require ('./envSwitcher');
-const {connectToRedis } = require('./redisClient');
+const envSwitcher = require ('../shared/envSwitcher');
+const {connectToRedis } = require('../shared/redisClient');
 
 let influxDbConfig;
 let influxDbClient;